split util file
