+++ pages/api/projects/join.ts
@@ -1,6 +1,6 @@
 import { NextApiRequest, NextApiResponse } from "next";
 import { getRequestUser } from "../../../utils/auth-utils";
-import { addUserToProject } from "../../../utils/mongo-utils";
+import { addUserToProject } from "../../../utils/mongo-users";
 
 export default async function handler(
   req: NextApiRequest,
